rename to ViRMENTemporalAlignmentBehaviorInterface
